separate github and discourse functions
